refactor (frontend): share useSigner across pages
